requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.12.0
lxml==5.1.0
selenium==4.16.0
webdriver-manager==4.0.1
